Fixed more bugs. Pending: calculating proper stack offsets.
It's easy enough to calculate the offset from the first env in the stack, however the problem is to keep track of the transition
offsets between environments: with recursive calls, the offset to the 'next' environment remains constant, but the following environment's offset
grows as the stack grows...

The environment-local offset is readily available. After it is accounted for, symbol lookups in the 'next' env must be offset by the amount of args in the function
that is being called (because we push args before we eval them and finally perform the call), plus the stack position of the return address.
Symbol lookups in the env following that must take that same offset, and add its own, similarly calculated offset.

This value is only resolvable at runtime since the stack grows dynamically in the case of recursion.

The basic concept is:
- track all push's and pop's in extra param in tail of env
- in every function call, insert an add of that number to %r15 (unused, won't interact with syscall's) before symbol resolution, subtract it before ret
- insert mov [local_offset + %r15](%rsp,n_args+1,8), [target register] for symbol resolution from stack positions

This will probably fail horribly due to some unforeseen corner case or three though.